[refactor] update template parameters and improve type handling in various classes, add executor argument in task.wait()

+++ include/ilias/platform/detail/iocp_overlapped.hpp
@@ -38,7 +38,7 @@ public:
      * 
      * @param offset 
      */
-    auto setOffset(size_t offset) -> void {
+    auto setOffset(uint64_t offset) -> void {
         ::ULARGE_INTEGER integer;
         integer.QuadPart = offset;
         Offset = integer.LowPart;
